+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.6 (2020-12-18)
+### Optimisations
+- syncPropertyView method from dom.js
+- rename OriginalObject into input in `inputHandler`
+- doc for `inputHandler` params
+
+### fixes
+- fix `radio` input views events
 ## 0.0.5 (2020-12-18)
 ### Optimisations
 - Prevent double mutations from DOM listeners